refactory: 카테고리, 태그 영역 공동 컴포넌트로 분리
